Applied renaming, updated implementation
