+++ server/vulnerabilities/nvd/cpe_matching_rules.go
@@ -1,0 +1,73 @@
+package nvd
+
+import (
+	"fmt"
+)
+
+type CPEMatchingRules []CPEMatchingRule
+
+func GetCPEMatchingRules() (CPEMatchingRules, error) {
+	// TODO: Move this to a metadata file?
+	rules := CPEMatchingRules{
+		CPEMatchingRule{
+			CPESpecs: []CPEMatchingRuleSpec{
+				{
+					Vendor:           "apple",
+					Product:          "icloud",
+					TargetSW:         "windows",
+					SemVerConstraint: "< 7.1",
+				},
+			},
+
+			CVEs: map[string]bool{
+				"CVE-2017-13797": true,
+			},
+		},
+		CPEMatchingRule{
+			CPESpecs: []CPEMatchingRuleSpec{
+				{
+					Vendor:           "apple",
+					Product:          "icloud",
+					TargetSW:         "windows",
+					SemVerConstraint: "<= 6.1.1",
+				},
+			},
+
+			CVEs: map[string]bool{
+				"CVE-2017-2383": true,
+			},
+		},
+		CPEMatchingRule{
+			CPESpecs: []CPEMatchingRuleSpec{
+				{
+					Vendor:           "apple",
+					Product:          "icloud",
+					TargetSW:         "windows",
+					SemVerConstraint: "<= 6.1.0",
+				},
+			},
+
+			CVEs: map[string]bool{
+				"CVE-2017-2366": true,
+			},
+		},
+	}
+
+	for i, rule := range rules {
+		if err := rule.Validate(); err != nil {
+			return nil, fmt.Errorf("invalid rule %d: %w", i, err)
+		}
+	}
+
+	return rules, nil
+}
+
+// FindMatch returns the first matching rule
+func (rules CPEMatchingRules) FindMatch(cve string) (*CPEMatchingRule, bool) {
+	for _, rule := range rules {
+		if rule.CVEs[cve] {
+			return &rule, true
+		}
+	}
+	return nil, false
+}
